@@ -3,26 +3,26 @@ import 'dart:collection';
 import 'package:flutter/material.dart';
 import 'package:flutter_html/flutter_html.dart';
 import 'package:flutter_html/html_parser.dart';
-import 'package:glow/external.dart';
-import 'package:glow/flyer.dart';
-import 'package:glow/resources.dart';
+import 'package:glow/External.dart';
+import 'package:glow/Flyer.dart';
+import 'package:glow/Resources.dart';
 import 'package:html/dom.dart' as dom;
 
-import 'common.dart';
+import 'Common.dart';
 
 class FlyerDetailPage extends StatefulWidget {
-  FlyerDetailPage({this.flyerId}) : super(key: new ValueKey(flyerId));
+  FlyerDetailPage(this.flyerId) : super(key: new ValueKey(flyerId));
   final String flyerId;
 
   @override
-  FlyerDetailPageState createState() =>
-      FlyerDetailPageState(flyer: Resources().flyerMap[flyerId]);
+  FlyerDetailPageState createState() {
+    return FlyerDetailPageState(Resources().flyerMap[flyerId]!);
+  }
 }
 
 class FlyerDetailPageState extends State<FlyerDetailPage> {
-  FlyerDetailPageState({this.flyer}) : super();
+  FlyerDetailPageState(this.flyer) : super();
   final Flyer flyer;
-  Html html;
   var top = 0.0;
   static const double BANNER_HEIGHT = 350;
 
@@ -51,7 +51,7 @@ class FlyerDetailPageState extends State<FlyerDetailPage> {
         future: flyer.loadContent(),
         builder: (BuildContext context, AsyncSnapshot<String> snapshot) {
           if (snapshot.hasData && snapshot.data != null) {
-            return getUiForLoaded(snapshot.data);
+            return getUiForLoaded(snapshot.data!);
           } else {
             return getUiForEmpty();
           }
@@ -64,12 +64,10 @@ class FlyerDetailPageState extends State<FlyerDetailPage> {
 
   Widget getUiForLoaded(String html) {
     return getContent(html);
-
   }
 
   Widget getContent(String html) {
-    return SingleChildScrollView(
-        child: getFlyerText(html));
+    return SingleChildScrollView(child: getFlyerText(html));
   }
 
   double getBannerHeight() {
@@ -81,21 +79,41 @@ class FlyerDetailPageState extends State<FlyerDetailPage> {
         padding: EdgeInsets.all(16),
         child: Html(
           data: html,
-          customRender: getCustomRender(),
-          onLinkTap: (link) => External.launchURL(link),
+          customImageRenders: getCustomImageRenders(),
+          onLinkTap: (url, context, attributes, element) => External.launchURL(url),
         ));
   }
 
+  Map<ImageSourceMatcher, ImageRender> getCustomImageRenders() {
+    var customRender = HashMap<ImageSourceMatcher, ImageRender>();
+    customRender[flyerAssetUriMatcher] = flyerAssetImageRender;
+    return customRender;
+  }
+
+  bool flyerAssetUriMatcher(Map<String, String> attributes, dom.Element? element) {
+    var src = attributes['src'];
+    return src != null ? src.contains('image') : false;
+  }
+
+  Widget? flyerAssetImageRender(
+    RenderContext context,
+    Map<String, String> attributes,
+    dom.Element? element,
+  ) {
+    return Padding(padding: EdgeInsets.all(24), child: new Image.asset("$FLYER${flyer.id}${attributes["src"]}"));
+  }
+
+/*
   Map<String, CustomRender> getCustomRender() {
     var customRender = HashMap<String, CustomRender>();
     customRender["img"] = getImageCustomRender;
     return customRender;
   }
 
-  Widget getImageCustomRender(RenderContext context, Widget parsedChild,
-      Map<String, String> attributes, dom.Element element) {
+  Widget getImageCustomRender(RenderContext context, Widget parsedChild) {
     return Padding(
         padding: EdgeInsets.all(24),
         child: new Image.asset("$FLYER${flyer.id}${attributes["src"]}"));
   }
+  */
 }
