@@ -1,6 +1,6 @@
 class FlyerMeta {
-  String title;
-  String url;
+  String title = "";
+  String url = "";
   FlyerMeta(Map<String, dynamic> data) {
     title = data['title'];
     url = data['url'];
